Added an option to use hard links for the first backup copy
When using ``rsync`` to make the first backup of the blobs, create hard
links instead of copies of the blobs.
This is useful to save disk space but you need to pay attention
that changing the blobs in the source folder
will change the blobs in the backup.
The ``blob_storage`` and the backup folder ``blobbackuplocation``
have to be in the same partition for hard links to be possible.

Fixes #57